Revert "Extend dynamicRef keyword"
It needs performance optimization. See python-jsonschema/jsonschema#941.

This reverts commit 12c791ee81baf27e54d18baf95b975bd48387dc6.
